+++ PatternPractice/DelegatePattern/PlusOneProtocol.swift
@@ -1,0 +1,14 @@
+//
+//  PlusOneProtocol.swift
+//  PatternPractice
+//
+//  Created by jeon sangjun on 2017/10/16.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+// プロトコルは「このメソッドやプロパティなどの要求事項で具現して」と命令。
+protocol PlusOneProtocol {
+    func resultOfPlusOne(value: Int)
+}
