+++ PatternPractice/FactoryPattern/Siege.swift
@@ -1,0 +1,15 @@
+//
+//  Siege.swift
+//  PatternPractice
+//
+//  Created by jeon sangjun on 2017/10/16.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+class Siege: Unit {
+    func action() {
+        print("시즈탱크가 포를 쏩니다.")
+    }
+}
