+++ PatternPractice/SingletonPattern/SingletonPattern.swift
