+++ PatternPractice/StrategyPattern/CharacterState.swift
@@ -1,0 +1,14 @@
+//
+//  File.swift
+//  PatternPractice
+//
+//  Created by jeon sangjun on 2017/10/16.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+protocol CharacterState {
+    func enterState()
+}
+
