+++ PatternPractice/StrategyPattern/StandingState.swift
@@ -1,0 +1,16 @@
+//
+//  Subtract.swift
+//  PatternPractice
+//
+//  Created by jeon sangjun on 2017/10/16.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+class StandingState: CharacterState {
+    func enterState() {
+        print("Entering Standing state")
+    }
+}
+
