+++ PatternPractice/StrategyPattern/WalkingState.swift
@@ -1,0 +1,16 @@
+//
+//  WalkingState.swift
+//  PatternPractice
+//
+//  Created by jeon sangjun on 2017/10/16.
+//  Copyright © 2017年 jeon sangjun. All rights reserved.
+//
+
+import Foundation
+
+class WalkingState: CharacterState {
+    func enterState() {
+        print("Entering Walking state")
+    }
+}
+
